@@ -31,17 +31,15 @@
   "devDependencies": {
     "@ndxbn/preset-jest": "0.2.4",
     "@ndxbn/preset-typescript": "0.0.12",
-    "@types/jest": "29.5.6",
-    "@types/node": "20.8.9",
-    "@typescript-eslint/eslint-plugin": "7.12.0",
-    "@typescript-eslint/parser": "7.12.0",
-    "eslint": "9.4.0",
+    "@types/jest": "29.5.12",
+    "@types/node": "22.5.4",
     "jest": "29.7.0",
-    "prettier": "3.0.3",
-    "ts-jest": "29.1.1",
-    "typescript": "5.2.2"
+    "prettier": "3.3.3",
+    "ts-jest": "29.2.5",
+    "ts-node": "10.9.2",
+    "typescript": "5.6.2"
   },
   "dependencies": {
-    "bson": "^6.0.0"
+    "bson": "6.8.0"
   }
 }
